+++ reactjs/src/components/SendComment.js
@@ -18,7 +18,7 @@ class SendComment extends AnyCommentComponent {
     render() {
         return (
             <div id="anycomment anycomment-send-comment"
-                 className={"anycomment-send-comment " + (!this.isGuest() ? 'anycomment-send-comment-authorized' : '') + ""}>
+                 className="anycomment-send-comment">
                 <div className="anycomment anycomment-send-comment-supheader">
                     <div className="anycomment anycomment-send-comment-supheader__count"
                          id="comment-count">{this.props.commentCountText}</div>