+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 **Enhancements:**
 
 * Added ability to rate page in comments as guest or logged in user, #156
-
+* Added ability to edit comment or reply to a comment right above it, #164
 
 **Fixes:**